pyproject.toml
[tool.poetry]
name = "zq-django-util"
version = "0.1.10"
description = "自强Studio Django 工具"
authors = ["Nagico <yjr888@vip.qq.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Nagico/zq-django-util"
documentation = "https://zq-django-util.readthedocs.io/en/latest/"
keywords = ["django", "drf", "util"]
packages = [{include = "zq_django_util"}]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
oss2 = ">=2.13.0,<3.0"
djangorestframework-simplejwt = ">=4.7,<6.0"
drf-standardized-errors = ">=0.9.0,<0.13.0"
django-cleanup = "^6.0.0"
isodate = "^0.6.1"

[tool.poetry.group.test.dependencies]
pytest = ">=6.2.0,<8.0"
pytest-cov = ">=4.0.0,<5.0"
pytest-django = ">=4.5.2,<5.0"
pytest-mock = ">=3.10.0,<4.0"
model-bakery = ">=1.9.0,<2.0.0"
django-filter = ">=22.1,<23.0"
sentry-sdk = "^1.9.10"

[tool.poetry.group.dev.dependencies]
django = ">=4.0,<5.0"
djangorestframework = ">=3.13,<4.0"
pytest = ">=6.2.0,<8.0"
pytest-cov = ">=4.0.0,<5.0"
pytest-django = ">=4.5.2,<5.0"
pytest-mock = ">=3.10.0,<4.0"
model-bakery = ">=1.9.0,<2.0.0"
importlib-metadata = "<6.0.0"
django-filter = ">=22.1,<23.0"
flake8-pyproject = "^1.2.2"
pre-commit = "^2.21.0"
sentry-sdk = "^1.9.10"
auto-changelog = "^0.6.0"
mkdocs = "^1.4.2"
mkdocstrings = {extras = ["python"], version = "^0.19.1"}
meilisearch = "^0.25.0"

[tool.poetry.extras]
sentry = ["sentry-sdk"]

[[tool.poetry.source]]
name = "tsinghua"
default = true
url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
skip = ["migrations"]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3
line_length = 80

# not using pytest-cov default, to support pycharm coverage
[tool.pytest.ini_options]
addopts = """
    -s -v
"""

[tool.coverage.run]
omit =[
  '*/migrations/*',
  '*/__version__.py',
  '*/tests/*',
  '*/types.py',
  '*/tests.py',
  '*/setup.py',
  '*/setup.cfg',
  '*/.tox/*',
  '*/.venv/*',
]

[tool.coverage.report]
exclude_lines =[
  'pragma: no cover',
  'def __repr__',
  'if TYPE_CHECKING',
  'raise AssertionError',
  'raise NotImplementedError',
  'if 0:',
  'if __name__ == .__main__.:',
  'class .*\bProtocol\):',
  '@(abc\.)?abstractmethod',
]

[tool.flake8]
ignore = ['E203', 'E266', 'E501', 'W503']
max-line-length = 80
max-complexity = 18
select = ['B', 'C', 'E', 'F', 'W', 'T4']
